+++ libs/bn-models/models/ashop/ashop-catch.ts
@@ -1,16 +1,22 @@
 // A-SHOP Catch
 import { BaseCatch } from '../_base/base-catch';
-import { Measurement, Basket } from '../_common/index';
-import { AshopSampleSpecies } from './ashop-sample-species';
-import { BaseSpecimen } from '../_base';
+import { Measurement, Basket, CouchID, Specimen } from '../_common/index';
 
-export const AshopSampleTypeName = 'ashop-sample';
+export const AshopCatchTypeName = 'ashop-catch';
 
 export interface AshopCatch extends BaseCatch {
+  // Catch = Sample in A-SHOP terminology, use sample for the UI
+
   sampleNum?: number; // unique/sequential within Haul
   isSubsample?: boolean; // for flattening convenience
   isPresorted?: boolean;
   isTruncated?: boolean; // Indicate a sample that was stopped (e.g. high species diversity)
+
+  isPredominantSpecies?: boolean;
+  isNotFlowScaleWeighed?: boolean; // not weighed by flow scale
+  percentRetained?: number;
+  sightingEventIds?: CouchID[];
+
   // species?: AshopSampleSpecies[];
   // subsamples?: AshopSample[];
 
@@ -22,9 +28,7 @@ export interface AshopCatch extends BaseCatch {
 
   children?: AshopCatch[];
 
-  baskets?: Basket[];
-  specimens?: BaseSpecimen[];
-
+  specimens?: Specimen[];
 
   legacy?: {
     sampleSequence?: number;
