+++ libs/bn-models/models/ashop/ashop-haul.ts
@@ -2,7 +2,7 @@
 import { LocationEvent, Measurement, CouchID } from '../_common/index';
 import { GearType } from '../_lookups/index';
 import { BaseOperation, BaseCatch } from '../_base/index';
-import { AshopSample } from './ashop-sample';
+import { AshopSample } from './ashop-catch';
 
 export const AshopHaulTypeName = 'ashop-haul';
