+++ libs/bn-models/models/ashop/index.ts
@@ -1,8 +1,5 @@
-export * from './ashop-basket';
-export * from './ashop-sample';
+export * from './ashop-catch';
 export * from './ashop-cruise';
 export * from './ashop-discard-reason';
 export * from './ashop-haul';
-export * from './ashop-sample-species';
-export * from './ashop-specimen';
 export * from './ashop-trip';
