+++ libs/bn-models/models/wcgop/index.ts
@@ -1,7 +1,6 @@
 export * from './wcgop-trip';
 export * from './wcgop-operation';
 export * from './wcgop-catch';
-export * from './wcgop-catch-species';
 export * from './wcgop-fish-ticket';
 export * from './wcgop-specimen';
 export * from './wcgop-discard-reason';
