+++ libs/bn-models/models/wcgop/wcgop-catch.ts
@@ -2,21 +2,32 @@
 // Import from CATCHES table
 
 import { BaseCatch } from '../_base/base-catch';
-import { Measurement, BoatnetDate, CouchID } from '../_common/index';
+import { Measurement, BoatnetDate, CouchID, Basket } from '../_common/index';
 import { WcgopCatchSpecies } from './wcgop-catch-species';
+import { WcgopDiscardReason } from './wcgop-discard-reason';
+import { WcgopSpecimen } from './wcgop-specimen';
 
 export const WcgopCatchTypeName = 'wcgop-catch';
+declare type RockfishHandlingCode = string; // TODO
 
 export interface WcgopCatch extends BaseCatch {
   catchNum?: number; // Unique per Operation sequential
   disposition?: string;
   weightMethod?: string;
-  species?: WcgopCatchSpecies[];
-  weight?: Measurement;
-  count?: number;
+  weight?: Measurement; // Could be species or higher level
+  count?: number; // Could be at species or higher level
   sampleWeight?: Measurement;
   sampleCount?: number;
   gearSegmentsSampled?: number;
+
+  // Species-level data
+  discardReason?: WcgopDiscardReason;
+  totalTally?: number;
+  handling?: RockfishHandlingCode; // Rockfish Handling
+  sightingEventIds?: CouchID[];
+  specimens?: WcgopSpecimen[];
+
+  children?: WcgopCatch[];
 
   legacy?: {
     catchId?: number;
@@ -41,5 +52,22 @@ export interface WcgopCatch extends BaseCatch {
 
     obsprodLoadDate?: BoatnetDate;
     hooksSampled?: number; // pulled up to Operation
+
+
+    // WcgopCatchSpecies - legacy items
+    speciesCompDataSource?: string;
+    speciesCompItemDataSource?: string;
+
+    speciesCompId?: number;
+    speciesCompItemId?: number;
+    biospecimenId?: number;
+    // catchId?: number;
+    speciesWeightKp?: number;
+    speciesWeightKpItq?: number;
+    speciesNumberKp?: number;
+    speciesNumberKpItq?: number;
+    catchSampleMethod?: string;
+    basketNumber?: number;
+    // obsprodLoadDate?: BoatnetDate;    
   };
 }
